requirements.txt
# Core Framework
fastapi>=0.127.1
uvicorn[standard]>=0.30.0
mcp>=0.1.0

# AI/ML Libraries
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
ollama>=0.1.6
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=4.0.0

# API & Web
requests>=2.31.0
duckduckgo-search>=6.0.0
python-multipart

# Database & Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL support
redis>=5.0.0  # For high-performance caching

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# PDF Processing
PyPDF2>=3.0.0

# Web Scraping & Research
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
ddgs>=0.1.0

# Code Quality & Formatting
black>=24.0.0
ruff>=0.1.0
isort>=5.13.0